requirements.txt
asgiref==3.9.1
cfgv==3.4.0
distlib==0.4.0
Django==5.2.6
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
identify==2.6.13
nodeenv==1.9.1
platformdirs==4.4.0
pre_commit==4.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
ruff==0.12.12
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
